+++ src/main/java/io/beanmapper/core/BeanFieldMatch.java
@@ -1,12 +1,15 @@
 package io.beanmapper.core;
 
 import java.lang.annotation.Annotation;
+import java.util.Map;
 
 import io.beanmapper.annotations.BeanDefault;
-import io.beanmapper.config.SecuredPropertyHandler;
+import io.beanmapper.annotations.LogicSecuredCheck;
+import io.beanmapper.config.RoleSecuredCheck;
 import io.beanmapper.core.converter.collections.BeanCollectionInstructions;
 import io.beanmapper.exceptions.BeanMappingException;
-import io.beanmapper.exceptions.BeanNoSecuredPropertyHandlerSetException;
+import io.beanmapper.exceptions.BeanNoLogicSecuredCheckSetException;
+import io.beanmapper.exceptions.BeanNoRoleSecuredCheckSetException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,35 +38,66 @@ public class BeanFieldMatch {
                 matchedBeanPairField.getSourceBeanField(),
                 matchedBeanPairField.getTargetBeanField());
     }
+
     public boolean hasAccess(
-            SecuredPropertyHandler securedPropertyHandler,
+            RoleSecuredCheck roleSecuredCheck,
+            Map<Class<? extends LogicSecuredCheck>, LogicSecuredCheck> logicSecuredChecks,
             Boolean enforcedSecuredProperties) {
 
-        if (securedPropertyHandler == null) {
+        boolean accessAllowed = checkForLogicSecured(
+                logicSecuredChecks, sourceBeanField, source, target, enforcedSecuredProperties);
+        accessAllowed = accessAllowed && checkForLogicSecured(
+                logicSecuredChecks, targetBeanField, source, target, enforcedSecuredProperties);
+
+        return accessAllowed && checkForRoleSecured(roleSecuredCheck, enforcedSecuredProperties);
+    }
+
+    private boolean checkForRoleSecured(RoleSecuredCheck roleSecuredCheck, Boolean enforcedSecuredProperties) {
+        if (roleSecuredCheck == null) {
             checkIfSecuredFieldHandlerNotSet(sourceBeanField, enforcedSecuredProperties);
             checkIfSecuredFieldHandlerNotSet(targetBeanField, enforcedSecuredProperties);
             return true;
         }
         return
-            securedPropertyHandler.hasRole(sourceBeanField.getRequiredRoles()) &&
-            securedPropertyHandler.hasRole(targetBeanField.getRequiredRoles());
+            roleSecuredCheck.hasRole(sourceBeanField.getRequiredRoles()) &&
+            roleSecuredCheck.hasRole(targetBeanField.getRequiredRoles());
+    }
+
+    private boolean checkForLogicSecured(
+            Map<Class<? extends LogicSecuredCheck>, LogicSecuredCheck> logicSecuredChecks,
+            BeanField beanField, Object source, Object target, Boolean enforcedSecuredProperties) {
+
+        Class<? extends LogicSecuredCheck> logicSecuredCheckClass = beanField.getLogicSecuredCheck();
+        if (logicSecuredCheckClass == null) {
+            return true;
+        }
+        LogicSecuredCheck logicSecuredCheck = logicSecuredChecks.get(beanField.getLogicSecuredCheck());
+        if (logicSecuredCheck == null) {
+            String message =
+                    "Property '" +
+                    beanField.getName() +
+                    "' has @BeanLogicSecured annotation, but bean for check is missing: " +
+                    logicSecuredCheckClass.getName();
+            if (enforcedSecuredProperties) {
+                throw new BeanNoLogicSecuredCheckSetException(message);
+            }
+            logger.warn(message);
+            return true;
+        }
+        return logicSecuredCheck.isAllowed(source, target);
     }
 
     private void checkIfSecuredFieldHandlerNotSet(BeanField beanField, Boolean enforcedSecuredProperties) {
         if (beanField.getRequiredRoles().length > 0) {
-            String message = getBeanFieldSecuredPropertyMessage(beanField);
+            String message =
+                    "Property '" +
+                    beanField.getName() +
+                    "' has @BeanRoleSecured annotation, but RoleSecuredCheck has not been set";
             if (enforcedSecuredProperties) {
-                throw new BeanNoSecuredPropertyHandlerSetException(message);
+                throw new BeanNoRoleSecuredCheckSetException(message);
             }
             logger.warn(message);
         }
-    }
-
-    private String getBeanFieldSecuredPropertyMessage(BeanField beanField) {
-        return
-                "Property '" +
-                beanField.getName() +
-                "' has @BeanSecuredProperty annotation, but SecuredPropertyHandler is not set";
     }
 
     public boolean hasSimilarClasses() {