+++ src/test/java/io/beanmapper/testmodel/beansecuredfield/CheckSameNameLogicCheck.java
@@ -1,0 +1,12 @@
+package io.beanmapper.testmodel.beansecuredfield;
+
+import io.beanmapper.annotations.LogicSecuredCheck;
+
+public class CheckSameNameLogicCheck implements LogicSecuredCheck<SFSourceELogicSecured, Object> {
+
+    @Override
+    public boolean isAllowed(SFSourceELogicSecured source, Object target) {
+        return "Henk".equals(source.name);
+    }
+
+}