+++ src/test/java/io/beanmapper/testmodel/beansecuredfield/SFTargetBWithSecuredField.java
@@ -1,10 +1,10 @@
 package io.beanmapper.testmodel.beansecuredfield;
 
-import io.beanmapper.annotations.BeanSecuredProperty;
+import io.beanmapper.annotations.BeanRoleSecured;
 
 public class SFTargetBWithSecuredField {
 
-    @BeanSecuredProperty({"ADMIN", "DEV"})
+    @BeanRoleSecured({"ADMIN", "DEV"})
     public String name;
 
 }